Refactor Android cloud repositories
